Split generator to components

+++ Source/MazeFirst/Generator/Cell.h
@@ -11,7 +11,7 @@ struct FCell
     
     uint8 X;
     uint8 Y;
-    uint32 Distance = 0;
+    uint16 Distance = 0;
     bool bHasLeft = true;
     bool bHasBottom = true;
     bool bIsVisited = false;
